+++ CHANGES.md
@@ -3,5 +3,6 @@ Gwtlib Changes
 
 master (top: more recent changes)
 ----------------------------------
+- added ListPanel and ListItemPanel
 - add SimpleMessageDialogBoxPresenter and MessageWithCommandDialogBoxPresenter
 - add CustomHTMLPanel and FireAndForgetCallback